% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_ex_var_scenario_2.R
\name{compare_ex_var_scenario_2}
\alias{compare_ex_var_scenario_2}
\title{Graphical displays of expectation or variance of different sampling schemes under scenario 2.}
\usage{
compare_ex_var_scenario_2(mulow, muhigh, sd, m1, m2, measure)
}
\arguments{
\item{mulow}{the lower value of the mean concentration (\eqn{\mu}) for use in the graphical display's x-axis.}

\item{muhigh}{the upper value of the mean concentration (\eqn{\mu}) for use in the graphical display's x-axis.}

\item{sd}{standard deviation on the log10 scale (default value 0.8).}

\item{m1}{the vector of the first set of incremental samples (with equal/unequal weights).}

\item{m2}{the vector of the second set of incremental samples (with equal/unequal weights).}

\item{measure}{what type of measure you would like to consider for the graph, such as "expectation" or "variance".}
}
\value{
Graphical displays based on expectation or variance when lot with heterogeneous and high-level contamination.
}
\description{
\code{\link{compare_ex_var_scenario_2}} provides graphical displays based on expectation or variance under scenario 2 of modelling the quantity of material sampled in the risk assessment study.
}
\details{
\code{\link{compare_ex_var_scenario_2}} provides graphical displays based on expectation or variance under scenario 2 of modelling the quantity of material sampled in the risk assessment study.
Under this scenario (a lot with homogeneous contaminations), we employed Poisson lognormal distribution to the model number of micro-organisms in the incremental samples. (this section will be updated later on)
}
\examples{
mulow <- 0
muhigh <- 2
m1 <- c(5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,
5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,
5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5)
m2 <- c(10,15,15,20,20,15,15,15,15,20,10,10,20,20,10,20,10,15,15,20,20,15,
15,15,15,20,10,10,20,20,10,20)
compare_ex_var_scenario_2(mulow, muhigh, sd = 0.8, m1, m2, measure = "variance")
compare_ex_var_scenario_2(mulow, muhigh, sd = 0.8, m1, m2, measure =  "expectation")
}
