% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_prevalence_scenario_4.R
\name{compare_prevalence_scenario_4}
\alias{compare_prevalence_scenario_4}
\title{Graphical displays of prevalence before inspection of different sampling schemes under scenario 4.}
\usage{
compare_prevalence_scenario_4(mulow, muhigh, sd, m1, m2, l, type, n_sim)
}
\arguments{
\item{mulow}{the lower value of the mean concentration (\eqn{\mu}) for use in the graphical display's x-axis.}

\item{muhigh}{the upper value of the mean concentration (\eqn{\mu}) for use in the graphical display's x-axis.}

\item{sd}{standard deviation on the log10 scale (default value 0.8).}

\item{m1}{the vector of the first set of incremental samples (with equal/unequal weights).}

\item{m2}{the vector of the second set of incremental samples (with equal/unequal weights).}

\item{l}{the number of lots in the production process.}

\item{type}{what type of the results you would like to consider such as "theory" or "simulation".}

\item{n_sim}{number of simulations (large simulations provide more precise estimation).}
}
\value{
Graphical displays based on prevalence before inspection when lot with homogeneous contamination and contamination levels fluctuate from lot to lot by using theoretical or simulation-based results.
}
\description{
\code{\link{compare_prevalence_scenario_4}} provides graphical displays based on prevalence before inspection under scenario 4 of modelling the quantity of material sampled in the risk assessment study.
}
\details{
\code{\link{compare_prevalence_scenario_4}} provides graphical displays based on prevalence before inspection under scenario 4 of modelling the quantity of material sampled in the risk assessment study. (this section will be updated later on)
}
\examples{
mulow <- -6
muhigh <- 1
m1 <- c(5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,
5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,
5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5)
m2 <- c(10,15,15,20,20,15,15,15,15,20,10,10,20,20,10,20,10,15,15,20,20,15,
15,15,15,20,10,10,20,20,10,20)
l <- 5000
compare_prevalence_scenario_4(mulow, muhigh, sd = 0.8, m1, m2, l, type = "theory")
}
