% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ex_var_scenario_2.R
\name{Ex_var_scenario_2}
\alias{Ex_var_scenario_2}
\title{Expected value or variance of microorganism in the aggregate sample under scenario 2.}
\usage{
Ex_var_scenario_2(mu, sd = 0.8, m, type, measure, n_sim)
}
\arguments{
\item{mu}{the the mean concentration (\eqn{\mu}).}

\item{sd}{standard deviation on the log10 scale (default value 0.8).}

\item{m}{the vector of incremental samples (with equal/unequal weights).}

\item{type}{what type of the results you would like to consider such as "theory" or "simulation".}

\item{measure}{what type of measure you would like to consider for the graph, such as "expectation" or "variance".}

\item{n_sim}{number of simulations (large simulations provide more precise estimation).}
}
\value{
expected value or variance of the number of microorganisms in the aggregate sample.
}
\description{
\code{\link{Ex_var_scenario_2}} provides the expected value or variance of the number of microorganisms in the aggregate sample under scenario 2 of modelling the quantity of material sampled in the risk assessment study.
}
\details{
\code{\link{Ex_var_scenario_2}} provides the expected value or variance of the number of microorganisms in the aggregate sample under scenario 2 of modelling the quantity of material sampled in the risk assessment study. (this section will be updated later on)
}
\examples{
mu <- 2
sd <- 0.8
m <- c(5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,
5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,
5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5)
Ex_var_scenario_2(mu, sd = 0.8, m, type = "theory",  measure = "expectation")
Ex_var_scenario_2(mu, sd = 0.8, m, type = "simulation",  measure = "expectation", n_sim = 2000)
Ex_var_scenario_2(mu, sd = 0.8, m, type = "theory",  measure = "variance")
Ex_var_scenario_2(mu, sd = 0.8, m, type = "simulation",  measure = "variance", n_sim = 2000)
}
