% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ex_var_scenario_3.R
\name{Ex_var_scenario_3}
\alias{Ex_var_scenario_3}
\title{Expected value or variance of microorganism in the aggregate sample under scenario 3.}
\usage{
Ex_var_scenario_3 (mu, sd, m, K, type, measure, n_sim)
}
\arguments{
\item{mu}{the the mean concentration (\eqn{\mu}).}

\item{sd}{standard deviation on the log10 scale (default value 0.8).}

\item{m}{the vector of incremental samples (with equal/unequal weights).}

\item{K}{shape parameter (default value 0.25).}

\item{type}{what type of the results you would like to consider such as "theory" or "simulation".}

\item{measure}{what type of measure you would like to consider for the graph, such as "expectation" or "variance".}

\item{n_sim}{number of simulations (large simulations provide more precise estimation).}
}
\value{
expected value or variance of the number of microorganisms in the aggregate sample.
}
\description{
\code{\link{Ex_var_scenario_3}} provides the expected value or variance of the number of microorganisms in the aggregate sample under scenario 3 of modelling the quantity of material sampled in the risk assessment study.
}
\details{
\code{\link{Ex_var_scenario_3}} provides the expected value or variance of the number of microorganisms in the aggregate sample under scenario 3 of modelling the quantity of material sampled in the risk assessment study. (this section will be updated later on)
}
\examples{
mu <- 2
sd <- 0.8
m <- c(5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,
5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,
5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5)
K <- 0.25
Ex_var_scenario_3(mu, sd = 0.8, m, K, type = "theory",  measure = "expectation")
Ex_var_scenario_3(mu, sd = 0.8, m, K, type = "simulation",  measure = "expectation",
 n_sim = 2000000)
Ex_var_scenario_3(mu, sd = 0.8, m, K, type = "theory",  measure = "variance")
Ex_var_scenario_3(mu, sd = 0.8, m, K, type = "simulation",  measure = "variance",
 n_sim = 2000000)
}
