% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_1_OC.R
\name{scenario_1_OC}
\alias{scenario_1_OC}
\title{Construction of  Operating Characteristic (OC) curve under lot with homogeneous contaminations.}
\usage{
scenario_1_OC(mulow, muhigh, sd, m, m1, n)
}
\arguments{
\item{mulow}{the lower value of the mean concentration (\eqn{\mu}) for use in the graphical display's x-axis.}

\item{muhigh}{the upper value of the mean concentration (\eqn{\mu}) for use in the graphical display's x-axis.}

\item{sd}{standard deviation on the log10 scale (default value 0.8).}

\item{m}{the quantity (weight) of the aggregate sample.}

\item{m1}{the quantity (weight) of the based sample for the risk assessment (for this research, we used a quantity of sample which is to be the minimum quantity of selected incremental samples).}

\item{n}{number of aggregate samples which are used for inspection.}
}
\value{
Operating Characteristic (OC) curves under lot with homogeneous contaminations.
}
\description{
\code{\link{scenario_2_OC}} provides the Operating Characteristic (OC) curves under scenario 1 of modelling the quantity of material sampled in the risk assessment study.
}
\details{
\code{\link{scenario_2_OC}} provides the Operating Characteristic (OC) curves under scenario 1 of modelling the quantity of material sampled in the risk assessment study.
Under this scenario (lot with homogeneous contaminations), we employed Poisson distribution to the model number of micro-organisms in the incremental samples.
The purpose of this function used for compares different sets of sampling schemes when lot with heterogeneous and high-level contamination.
Under this scenario expected cell count in each incremental sample can be written in terms of the based incremental sample's expected cell count (for this study, the based sample is a sample that is to be the minimum quantity).
The probability of acceptance is plotted against mean log10 concentration and expected cell counts. Based on the food safety literature, the expected cell count is given by \eqn{\lambda = 10^{\mu+log(10)\sigma^2/2}}. (this section will be updated later on)
}
\examples{
mulow <- -6
muhigh <- 0
sd <- 0.8
m <- 25
m1 <- c(5,10,15)
n <- 10
scenario_1_OC(mulow, muhigh, sd, m, m1, n)
}
\seealso{
\link{scenario_1_pd}
}
