% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_1_pa.R
\name{scenario_1_pa}
\alias{scenario_1_pa}
\title{Probability of acceptance estimation when lot with homogeneous contaminations.}
\usage{
scenario_1_pa(c, mu, sd, m, n, type, n_sim)
}
\arguments{
\item{c}{acceptance number}

\item{mu}{the the mean concentration (\eqn{\mu}).}

\item{sd}{standard deviation on the log10 scale (default value 0.8).}

\item{m}{the vector of incremental samples (with equal/unequal weights).}

\item{n}{number of aggregate samples which are used for inspection.}

\item{type}{what type of the results you would like to consider such as "theory" or "simulation".}

\item{n_sim}{number of simulations (large simulations provide more precise estimation).}
}
\value{
Probability of acceptance under lot with homogeneous contaminations.
}
\description{
\code{\link{scenario_1_pa}} provides a probability of acceptance under scenario 1 of modelling the quantity of material sampled in the risk assessment study.
}
\details{
\code{\link{scenario_1_pa}} provides a probability of acceptance under scenario 1 of modelling the quantity of material sampled in the risk assessment study.
Under this scenario (a lot with homogeneous contaminations), we employed Poisson distribution to the model number of micro-organisms in the incremental samples. Based on the food safety literature, the expected cell count is given by \eqn{\lambda = 10^{\mu+log(10)\sigma^2/2}}. (this section will be updated later on)
}
\examples{
c <- 0
mu <- -3
sd <- 0.8
m <- c(5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,
5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,
5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5)
n <- 10
scenario_1_pa(c, mu, sd = 0.8, m, n, type = "theory")
scenario_1_pa(c, mu, sd = 0.8, m, n, type = "simulation", n_sim = 2000000)
}
