% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_1_pd.R
\name{scenario_1_pd}
\alias{scenario_1_pd}
\title{Probability of detection estimation when lot with homogeneous contaminations.}
\usage{
scenario_1_pd(lambda, M, m1)
}
\arguments{
\item{lambda}{expected cell count in the based sample (for this research, we used a quantity of sample which is to be the minimum quantity of selected incremental samples).}

\item{M}{the quantity (weight) of the aggregate sample.}

\item{m1}{the quantity (weight) of the based sample for the risk assessment.}
}
\value{
Probability of detection under lot with homogeneous contaminations.
}
\description{
\code{\link{scenario_1_pd}} provides a probability of detection under scenario 1 of modelling the quantity of material sampled in the risk assessment study.
}
\details{
\code{\link{scenario_1_pd}} provides a probability of detection under scenario 1 of modelling the quantity of material sampled in the risk assessment study.
Under this scenario (a lot with homogeneous contaminations), we employed Poisson distribution to the model number of micro-organisms in the incremental samples. Based on the food safety literature, the expected cell count is given by \eqn{\lambda = 10^{\mu+log(10)\sigma^2/2}}. (this section will be updated later on)
}
\examples{
lambda <- 0.05
M <- 60
m1 <- 5
scenario_1_pd(lambda, M, m1)
}
