% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_2_pa.R
\name{scenario_2_pa}
\alias{scenario_2_pa}
\title{Probability of acceptance estimation when lot with heterogeneous and high-level contamination.}
\usage{
scenario_2_pa(c, mu, sd, m, n, type, n_sim)
}
\arguments{
\item{c}{acceptance number}

\item{mu}{the the mean concentration (\eqn{\mu}).}

\item{sd}{standard deviation on the log10 scale (default value 0.8).}

\item{m}{the vector of incremental samples (with equal/unequal weights).}

\item{n}{number of aggregate samples which are used for inspection.}

\item{type}{what type of the results you would like to consider such as "theory" or "simulation".But,not yet established for this scenario at this time.}

\item{n_sim}{number of simulations (large simulations provide more precise estimation).}
}
\value{
Probability of acceptance when lot with heterogeneous and high-level contamination.
}
\description{
\code{\link{scenario_2_pa}} provides a probability of acceptance under scenario 2 of modelling the quantity of material sampled in the risk assessment study.
}
\details{
\code{\link{scenario_2_pa}} provides a probability of acceptance under scenario 2 of modelling the quantity of material sampled in the risk assessment study.
Under this scenario (a lot with heterogeneous, high-level contamination), we employed Poisson lognormal distribution to the model number of micro-organisms in the incremental samples. Based on the food safety literature, the expected cell count is given by \eqn{\lambda = 10^{\mu+log(10)\sigma^2/2}}. (this section will be updated later on)
}
\examples{
c <- 0
mu <- -3
sd <- 0.8
m <- c(5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,
5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,
5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5)
n <- 10
scenario_2_pa(c, mu, sd = 0.8, m, n, type = "theory")
scenario_2_pa(c, mu, sd = 0.8, m, n, type = "simulation", n_sim = 200000)
}
