% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_2_pd_curve.R
\name{scenario_2_pd_curve}
\alias{scenario_2_pd_curve}
\title{Construction of  Operating Characteristic (OC) curve under lot with heterogeneous and high-level contamination.}
\usage{
scenario_2_pd_curve(mulow, muhigh, sd, m1, m2, type, n_sim)
}
\arguments{
\item{mulow}{the lower value of the mean concentration (\eqn{\mu}) for use in the graphical display's x-axis.}

\item{muhigh}{the upper value of the mean concentration (\eqn{\mu}) for use in the graphical display's x-axis.}

\item{sd}{standard deviation on the log10 scale (default value 0.8).}

\item{m1}{the vector of the first set of incremental samples (with equal/unequal weights).}

\item{m2}{the vector of the second set of incremental samples (with equal/unequal weights).}

\item{type}{what type of the results you would like to consider such as "theory" or "simulation".}

\item{n_sim}{number of simulations (large simulations provide more precise estimation).}
}
\value{
probability of detection curves when lot with heterogeneous and high-level contamination.
}
\description{
\code{\link{scenario_2_pd_curve}} provides the probability of detection curves under scenario 2 of modelling the quantity of material sampled in the risk assessment study.
}
\details{
\code{\link{scenario_2_pd_curve}} provides the probability of detection curves under scenario 2 of modelling the quantity of material sampled in the risk assessment study. (this section will be updated later on)
}
\examples{
mulow <- -10
muhigh <- 0
sd <- 0.8
m1 <- c(5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,
5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,
5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5)
m2 <- c(10,15,15,20,20,15,15,15,15,20,10,10,20,20,10,20,10,15,15,20,20,15,
15,15,15,20,10,10,20,20,10,20)
scenario_2_pd_curve(mulow, muhigh, sd = 0.8, m1, m2, type = "simulation", n_sim = 20)
}
\seealso{
\link{scenario_2_pd}
}
