% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_3_pd.R
\name{scenario_3_pd}
\alias{scenario_3_pd}
\title{Probability of detection estimation when lot with heterogeneous and low-level contamination.}
\usage{
scenario_3_pd(mu, sd, m, K, n_sim)
}
\arguments{
\item{mu}{the the mean concentration (\eqn{\mu}).}

\item{sd}{standard deviation on the log10 scale (default value 0.8).}

\item{m}{the vector of incremental samples(with equal/unequal weights).}

\item{K}{shape parameter (default value 0.25).}

\item{n_sim}{number of simulations (large simulations provide more precious estimation).}
}
\value{
Probability of detection when lot with heterogeneous and low-level contamination.
}
\description{
\code{\link{scenario_3_pd}} provides a probability of detection under scenario 2 of modelling the quantity of material sampled in the risk assessment study.
}
\details{
\code{\link{scenario_3_pd}} provides a probability of detection under scenario 2 of modelling the quantity of material sampled in the risk assessment study.
Under this scenario (a lot with heterogeneous, low-level contamination), we employed Poisson gamma distribution to the model number of micro-organisms in the incremental samples. Based on the food safety literature, the expected cell count is given by \eqn{\lambda = 10^{\mu+log(10)\sigma^2/2}}. (this section will be updated later on)
}
\examples{
mu <- -3
sd <- 0.8
m <- c(5,10,15,5,10,10,5)
K <- 0.05
n_sim <- 100000
scenario_3_pd(mu, sd, m, K, n_sim)
}
