% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_5_pd_curve.R
\name{scenario_5_pd_curve}
\alias{scenario_5_pd_curve}
\title{Construction of   probability of detection curves under lot with heterogeneous contamination and concentration levels fluctuating from sublot to sublot.}
\usage{
scenario_5_pd_curve(mulow, muhigh, sd_b, sd_w, m11, m22, type, n_sim)
}
\arguments{
\item{mulow}{the lower value of the mean concentration (\eqn{\mu}) for use in the graphical display's x-axis.}

\item{muhigh}{the upper value of the mean concentration (\eqn{\mu}) for use in the graphical display's x-axis.}

\item{sd_b}{standard deviation of concentration level between sublots on the log10 scale (default value 0.8).}

\item{sd_w}{standard deviation of concentration level within sublot on the log10 scale (default value 0.8).}

\item{m11}{the vector of the first set of incremental samples (with equal/unequal weights).}

\item{m22}{the vector of the second set of incremental samples (with equal/unequal weights).}

\item{type}{what type of the results you would like to consider such as "theory" or "simulation".}

\item{n_sim}{number of simulations (large simulations provide more precise estimation).}
}
\value{
probability of detection curves when lot with heterogeneous and high-level contamination.
}
\description{
\code{\link{scenario_5_pd_curve}} provides the probability of detection curves under scenario 5 of modelling the quantity of material sampled in the risk assessment study.
}
\details{
\code{\link{scenario_5_pd_curve}} provides the probability of detection curves under scenario 5 of modelling the quantity of material sampled in the risk assessment study. (this section will be updated later on)
}
\examples{
m1 <- c(10,10,10,10,10,10,10,10,10,10)
m2 <- c(10,10,10,10,10,10,10,10,10,10)
m3 <- c(10,10,10,10,10,10,10,10,10,10)
m4 <- c(10,10,10,10,10,10,10,10,10,10)
m5 <- c(10,10,10,10,10,10,10,10,10,10)
m11 <- list(m1,m2,m3,m4,m5)
m_1 <- c(15, 5, 5, 5, 10, 5, 10, 5, 15, 10)
m_2 <- c(5, 10, 5, 25, 10, 5, 10, 5, 5, 10)
m_3 <- c(5, 15, 10, 5, 5, 20, 5, 10, 5, 10)
m_4 <- c(20, 5, 10, 30, 5, 20, 5, 10, 5, 10)
m_5 <- c(20, 15, 10, 15, 10, 10, 5, 10, 15, 5)
m22 <- list(m_1,m_2,m_3,m_4,m_5)
mulow <- -8
muhigh <- -1
sd_b <- 0.2
sd_w <- 0.8
scenario_5_pd_curve(mulow, muhigh, sd_b, sd_w, m11, m22, type = "theory")
scenario_5_pd_curve(mulow, muhigh, sd_b, sd_w, m11, m22, type = "simulation", n_sim = 1000000)
}
\seealso{
\link{scenario_5_pd}
}
