% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarious_1_OC.R
\name{scenarious_1_OC}
\alias{scenarious_1_OC}
\title{Construction of  Operating Characteristic (OC) curve under lot with homogeneous contaminations.}
\usage{
scenarious_1_OC(mulow,muhigh,sd,m,m1,n)
}
\arguments{
\item{mulow}{the lower value of the mean concentration (\eqn{\mu}) for use in the graphical display's x-axis.}

\item{muhigh}{the upper value of the mean concentration (\eqn{\mu}) for use in the graphical display's x-axis.}

\item{sd}{standard deviation of the lognormal and Poisson-lognormal distributions on the log10 scale (default value 0.8)}

\item{m}{the quantity (weight) of the aggregate sample}

\item{m1}{the quantity (weight) of the based sample for the risk assessment (for this research, we used a quantity of sample which is to be the minimum quantity of selected incremental samples)}

\item{n}{number of aggregate samples which are used for inspection}
}
\value{
Operating Characteristic (OC) curves under lot with homogeneous contaminations
}
\description{
\code{\link{scenarious_1_OC}} provides the Operating Characteristic (OC) curve for known microbiological distribution such as Poisson.
The probability of acceptance is plotted against mean log10 concentration and expected cell counts.
}
\details{
Based on the food safety literature, the expected cell count is given by \eqn{\lambda = 10^{\mu+log(10)\sigma^2/2}}. (this section will be updated soon)
}
\examples{
mulow <- -6
muhigh <- 0
sd <- 0.8
m <- 25
m1 <- c(5,10,15)
n <- 20
scenarious_1_OC(mulow,muhigh,sd,m,m1,n)
}
