% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AOQL_scenarios.R
\name{AOQL_scenarios}
\alias{AOQL_scenarios}
\title{Construction of  AOQ curve and calculate AOQL value based on average microbial counts}
\usage{
AOQL_scenarios(c,llim, sd, m, n, scenario, K, n_sim)
}
\arguments{
\item{c}{acceptance number}

\item{llim}{the upper limit for graphing the arithmetic mean of cell count}

\item{sd}{standard deviation on the log10 scale (default value 0.8).}

\item{m}{the vector of incremental samples (with equal/unequal weights).}

\item{n}{number of aggregate samples which are used for inspection.}

\item{scenario}{what scenario we have considered such as  \code{"1B"}or \code{"2"}or \code{"3"}}

\item{K}{dispersion parameter of the Poisson gamma distribution (default value 0.25)}

\item{n_sim}{number of simulations (large simulations provide more precise estimation).}
}
\value{
AOQ curve and AOQL value based on expected microbial counts in each scenario.
}
\description{
\code{\link{AOQL_scenarios}} provides the Average Outgoing Quality (AOQ) curve and calculates Average Outgoing Quality Level (AOQL) value based on expected microbial counts in each scenario.
}
\details{
Since \eqn{P_a} is the probability of acceptance, \eqn{\lambda} is the arithmetic mean of cell count and the outgoing contaminated arithmetic mean of cell count of incremental samples is given by \eqn{AOQ} as the product \eqn{\lambda P_a}.
          The quantity \eqn{AOQL} is defined as the maximum proportion of outgoing contaminated incremental samples and is given by \deqn{AOQL ={\max_{\lambda \geq 0}}{\lambda P_a}}
}
\examples{
c <- 0
llim <- 0.5
sd <- 0.8
m1 <- c(10,10,10,10,10,10)
m2 <- c(10,12,18,20)
n <- 10
scenario <- "1B"
n_sim <- 100000
AOQL_scenarios(c,llim, sd, m=m1, n, scenario, K, n_sim)
AOQL_scenarios(c,llim, sd, m=m2, n, scenario, K, n_sim)
}
\seealso{
\link{scenario_1B_pa}, \link{scenario_2_pa},\link{scenario_3_pa}
}
