% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_3_OC.R
\name{scenario_3_OC}
\alias{scenario_3_OC}
\title{Construction of  Operating Characteristic (OC) curve under lot with heterogeneous and low-level contamination.}
\usage{
scenario_3_OC(mulow, muhigh, sd, m1, m2, K, n, n_sim)
}
\arguments{
\item{mulow}{the lower value of the mean concentration (\eqn{\mu}) for use in the graphical display's x-axis.}

\item{muhigh}{the upper value of the mean concentration (\eqn{\mu}) for use in the graphical display's x-axis.}

\item{sd}{standard deviation on the log10 scale (default value 0.8).}

\item{m1}{the vector of the first set of incremental samples (with equal/unequal weights).}

\item{m2}{the vector of the second set of incremental samples (with equal/unequal weights).}

\item{K}{shape parameter (default value 0.25).}

\item{n}{number of aggregate samples which are used for inspection.}

\item{n_sim}{number of simulations (large simulations provide more precious estimation).}
}
\value{
Operating Characteristic (OC) curves when lot with heterogeneous and low-level contamination.
}
\description{
\code{\link{scenario_3_OC}} provides the Operating Characteristic (OC) curves under scenario 2 of modelling the quantity of material sampled in the risk assessment study.
}
\details{
\code{\link{scenario_3_OC}} provides the Operating Characteristic (OC) curves under scenario 2 of modelling the quantity of material sampled in the risk assessment study.
The purpose of this function used for compares two different sets of sampling schemes when lot with heterogeneous and low-level contamination.
Nevertheless, each sampling scheme's total quantity (weight of aggregate sample (say M)) must be equal.
We employed Poisson gamma distribution to the model number of micro-organisms in the incremental samples. Based on the food safety literature, the expected cell count is given by \eqn{\lambda = 10^{\mu+log(10)\sigma^2/2}}. (this section will be updated later on)
}
\examples{
mulow <- -5
muhigh <- 2
sd <- 0.8
m1 <- c(10,10,10,10,10,10)
m2 <- c(10,12,18,20)
K <- 0.05
n <- 10
n_sim <- 100000
scenario_3_OC(mulow, muhigh, sd, m1, m2, K, n, n_sim)
}
\seealso{
\link{scenario_3_pd}
}
